move dict to separate file
