+++ src/dict.ts
@@ -1,0 +1,7 @@
+export const dict = {
+	appName: 'Pulse Remote',
+	headerInput: 'Pulse Remote: Input',
+	headerOutput: 'Pulse Remote: Output',
+	headerAbout: 'Pulse Remote: About',
+	headerConfig: 'Pulse Remote: Config',
+} as const
