+++ src/pages/controller-output.tsx
@@ -2,7 +2,7 @@ import { Fragment } from 'react'
 import { useVolumeStatus } from '../api/use-vol-status'
 import { Layout } from '../components/layout'
 import { VolumeSlider } from '../components/volume-slider'
-import { dict } from '../constant'
+import { dict } from '../dict'
 
 export const ControllerOutput: React.FC = () => {
 	const vol = useVolumeStatus()
